+++ exolve-from-puz.js
@@ -24,7 +24,7 @@ SOFTWARE.
 The latest code and documentation for Exolve can be found at:
 https://github.com/viresh-ratnakar/exolve
 
-Version: Exolve v1.34 March 20, 2022
+Version: Exolve v1.35 March 20, 2022
 */
 
 function exolveFromPuzNextNull(buffer, offset) {